+++ docs/changelog.md
@@ -21,6 +21,7 @@ Added:
 
 - Add delete button for activities (with an “Are you sure?” confirmation). ([GH-278](https://github.com/martin-ueding/geo-activity-playground/issues/278))
 - Add user defined colors for activity tags. ([GH-276](https://github.com/martin-ueding/geo-activity-playground/issues/276))
+- Add download of original file (after login). ([GH-271](https://github.com/martin-ueding/geo-activity-playground/issues/271))
 
 ## Version 0.43.3 — 2025-05-15